Add initial functions for the harvester api
