+++ pkg/harvester/api/api.go
@@ -36,68 +36,3 @@ func (a *HTTPApi) Run(ctx context.Context) error {
 	<-ctx.Done()
 	return nil
 }
-
-func (a *HTTPApi) GetTrustBundle(ctx context.Context, spiffeID string) (common.TrustBundle, error) {
-	telemetry.Count(ctx, telemetry.HTTPApi, telemetry.TrustBundle, telemetry.Get)
-
-	var tb common.TrustBundle
-
-	tb, err := a.controller.GetTrustBundle(ctx, spiffeID)
-	if err != nil {
-		a.logger.Error(err)
-		return tb, err
-	}
-
-	return tb, nil
-}
-
-func (a *HTTPApi) AddTrustBundle(ctx context.Context, trustBundle common.TrustBundle) (common.TrustBundle, error) {
-	telemetry.Count(ctx, telemetry.HTTPApi, telemetry.TrustBundle, telemetry.Add)
-
-	tb, err := a.controller.AddTrustBundle(ctx, trustBundle)
-	if err != nil {
-		a.logger.Error(err)
-		return tb, err
-	}
-
-	return tb, nil
-}
-
-// POST: federation-relationship/{relationshipId} {action: approve/deny}
-func (a *HTTPApi) ManageFederationRelationship(ctx context.Context, id string) (common.FederationRelationship, error) {
-	var fr common.FederationRelationship
-
-	// if body.action == telemetry.Approve {
-	// 	telemetry.Count(ctx, telemetry.HTTPApi, telemetry.TrustBundle, telemetry.Approve)
-	// 	fr, err := a.controller.ApproveFederationRelationship(ctx, id)
-	//  if err != nil {
-	//    a.logger.Error(err)
-	//    return fr, err
-	//   }
-	// }
-
-	// if body.action == telemetry.Deny {
-	// 	telemetry.Count(ctx, telemetry.HTTPApi, telemetry.TrustBundle, telemetry.Approve)
-	//  fr, err := a.controller.DenyFederationRelationship(ctx, id)
-	//  if err != nil {
-	//    a.logger.Error(err)
-	//    return fr, err
-	//  }
-	// }
-
-	return fr, nil
-}
-
-func (a *HTTPApi) GetFederationRelationshipsByStatus(ctx context.Context, status string) ([]common.FederationRelationship, error) {
-	telemetry.Count(ctx, telemetry.HTTPApi, telemetry.FederationRelationship, telemetry.Get)
-
-	var fr []common.FederationRelationship
-
-	fr, err := a.controller.GetFederationRelationshipsByStatus(ctx, status)
-	if err != nil {
-		a.logger.Error(err)
-		return fr, err
-	}
-
-	return fr, nil
-}
